+++ llmchain/loaders/tests/it/main.rs
@@ -12,5 +12,5 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+mod document_splitter;
 mod markdowns;
-mod splitter;
